refactor and documentation
